BaseModel toegevoegd

+++ src/main/java/nl/novi/kapsalon/repositories/TreatmentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface TreatmentRepository extends JpaRepository<Treatment, Integer> {
-    List<Treatment> findAllByIdIn(List<Integer> treatmentIds);
+public interface TreatmentRepository extends JpaRepository<Treatment, Long> {
+    List<Treatment> findAllByIdIn(List<Long> treatmentIds);
 }
